band create

+++ app/Http/Controllers/Profile/GalleryController.php
@@ -7,12 +7,15 @@ use App\Http\Requests\Gallery\GalleryUpdateRequest;
 use App\Models\Gallery;
 use App\Repositories\BandRepository;
 use App\Services\GalleryService;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Illuminate\Http\RedirectResponse;
 use Inertia\Inertia;
 use Inertia\Response;
 
 class GalleryController
 {
+    use AuthorizesRequests;
+
     public function __construct(protected GalleryService $galleryService)
     {
     }
@@ -35,6 +38,8 @@ class GalleryController
 
     public function edit(Gallery $gallery): Response
     {
+        $this->authorize('update', $gallery);
+
         return Inertia::render('Profile/Gallery/GalleryCreateEdit', [
             'gallery' => $gallery->load('venue'),
             'bandsList' => BandRepository::bandList(),
@@ -43,6 +48,8 @@ class GalleryController
 
     public function update(GalleryUpdateRequest $request, Gallery $gallery): RedirectResponse
     {
+        $this->authorize('update', $gallery);
+
         $this->galleryService->update($gallery, $request->validated());
 
         session()->flash('message', 'The gallery has been updated.');
@@ -52,6 +59,8 @@ class GalleryController
 
     public function destroy(Gallery $gallery): RedirectResponse
     {
+        $this->authorize('delete', $gallery);
+
         $this->galleryService->destroy($gallery);
 
         session()->flash('message', 'The gallery has been deleted.');
